[qwaq] Show the quadsphere
Unfortunately, all lod levels are shown at once so only the base cube is
visible, but the meshes show nicely in renderdoc's mesh viewer, and I
got that really nice triangle pattern I was wondering how to create...
without even trying. I guess it's a natural consequence of using
half-edges and the algorithm used for assigning new verts and edges.