[qfcc] Disallow shape-changing casts for inout params
inout is *NOT* a reference, so it will implicity alter values when
passed to a function even if the function does nothing.  Apprarently
GLSL does allow this, but it's considered to be bad (and I do not
disagree).  Aliases are bit-casts and they're considered to be ok
because there's no loss of data. Thanks to devsh for making me aware of
the issue.