[qwaq] Add a little test for building a leaf
The leaf is extracted from ad_tears (actually, the very last visLEaf in
the map). The brush gizmo came in handy for seeing what the leaf looks
like (after shrinking it down to qwaq scale), and the capsule gizmos for
seeing the edges I had extracted. However it turns out that particular
approach to finding a point in the leaf's volume won't work (most of the
lines are outside the volume, so I guess it was a good choice for
testing). Still, as a first attempt, it was a nice test of using GA for
finding those lines and points.