[qfcc] Use skip marker in merge_extends
Moving merge_extends to expr_optimize is because it made more sense to
expose the function than to expose the skip marker. This fixes losing
the whole expression when the first (of more than one) add is canceled
by a sub, which also fixes the broken type for `~To * P * To` in
leafnode.r.