[thread] Add a job system
The implementation is based on:
 * An Efficient Work-Stealing Scheduler for Task Dependency Graph
   Chun-Xun Lin, Tsung-Wei Huang, Martin D. F. Wong
   https://tsung-wei-huang.github.io/papers/icpads20.pdf
 * Dynamic Circular Work-Stealing Deque
   David Chase, Yossi Lev
   https://www.dre.vanderbilt.edu/~schmidt/PDF/work-stealing-dequeue.pdf
 * The notifier class in the Eigen library.

While I had tested the code pretty thoroughly on x86 (linux and windows)
and arm (linux), it turns out there's still a bug where DEQUE_ABORT is
getting returned as a task when I think it shouldn't be. I hope I can
get this working.