[model] Implement quickhull
I think this is the 4th time I've implemented quickhull, but it looks
like this might be the fastest version (though when I add point-on-edge
checks, it might slow down a bit). The test fails because I'm not sure
just what to check for other than outright failure (which needs some
test data that fails). Also, something to keep in mind is the faces are
constructed using counter-clockwise winding (from outside), which is the
reverse of quake. This is a big step towards being able to robustly
tessellate BSP models.