[qfcc] Add some comments about the qv implementation
Notably that it will not give correct results if the quaternion isn't
normalized and why. I might add an attribute that disables the
optimization (or go the other way and the attribute allows it), but it
is nice to have a note as to why it works: I was rather surprised to
find almost half the math missing when I compared it to what qfcc emits
for the GA equivalent. I also want to add a check for `a×(b×c) + a•c b`
and replace it with `2*a×(b×c) + a•b c`: even without the normalized
assumption, it looks like it can reduce the instruction count by one for
`~q * v * q` in 3d VGA.