[qfcc] Fix rotating a VGA mirror
Well, actually, just doing the reverse first on the quaternion when
rotating a vector. At first, it wouldn't compile due to the sign of the
reversed bivector being in the wrong place in the expression dag, then
the resulting code was broken due to misidentifying `a×(b×c) + a•b` (had
b and c swapped in the test).