[set] Add set_start for arbitrary iteration start points
In the process of rewriting the scrap allocator, I needed a nice way to
start iterating at a certain point. While it could have been achieved
via set_first then setting the iterator's element, that felt rather
wrong (element is meant to be immutable, but C's not super helpful
there).