[renderer] Rework the scrap allocator to not be O(N)
There is some variation as the number of allocations grows, but it's no
longer O(N) (N = number of existing allocations), so ad_tears now takes
about 0.5s instead of about 40s to load. There's some odd behavior with
allocations for the 2d sprites, but lightmaps and fonts allocate nicely
(still get nice horizontal bars).

Merging when freeing rects is disabled for now: I made some progress on
it, but I realized I don't have good tests for it (it's used only during
shutdown normally, so it's not a huge deal yet).