[renderer] Clear size counts when clearing scrap
Fixes the full size free width/height not getting removed for a
previously cleared scrap. Found when trying to load ad_tears in the
forward/lightmap vulkan renderer (which clears the scrap before first
use, which gl and glsl don't do). There are still problems with the
search algorithm as allocation fails for a 2k scrap when it shouldn't
(but a 4k scrap is fine).