[vulkan] Separate dynamic descriptor alloc/update
Frames are split into two parts:
```
--3d--|--2d--||--3d--|--2d--||...
      a      b       a      b
<--frame 0-->  <--frame 1-->
```

Where `a` and `b` are fences. At the very beginning of the render graph
for frame 2, `wait_on_fence` waits on fence `a` from frame 0, followed
by a few other functions including `draw_scr_funcs`, which takes care of
collecting all the 2d commands to be submitted in the 2d segment. This
included updating any descriptors as needed.

The problem is because `wait_on_fence` waits on fence a, it's possible
for `draw_scr_funcs` in frame 2 to execute before frame 0's 2d rendering
has completed. However, `acquire_output` waits on fence `b` and is
called before any actual 2d rendering is started (via `flush_draw`).

Thus, record the necessary information for the dynamic descriptor sets
when collecting the 2d commands (in `draw_scr_funcs`), but actually
update the descriptor sets (using the recorded information) just before
rendering the 2d (in `flush_draw`). This seems to work very nicely as
the validation error became very reliable with the entity culling in a
smashed state and running an optimized build, but has disappeared with
this change.