[qfcc] Add a failing test for aliased assignments
This was a nasty one to reproduce, which is probably why it happened
only in qwaq-app and only in Debugger's initWithTarget (and thus I
missed it until I tried running qdb).

The nature of the problem is s gets assigned to the extent member of the
temporary rect that gets passed to set_rect ([Window withRect] in
qwaq-app) and then when coordinates are calculated (via the subtraction)
for create ([ScrollBar vertica:at:] in qwaq-app), the temporary rect is
reused instead of s. This in itself is not a problem but rather the
offset of the extent member is calculated incorrectly (or more likely
ignored) resulting in the position member being used.

At first I thought 07158c4cc4494af297d2708469dbcfce91ccb655 was the
cause of the problem, but that was before I'd noticed the mangled
offsets. Rather, that change exposed the bug because the aggressive
killing of nodes gave the bug somewhere to hide due to forcing the reuse
of s instead of the temporary rect's extent member.