[vulkan] Fix bsp texture loading for ad_tears
The recent changes to texture loading means that care has to be taken
with mipmaps and barriers, and ad_tears has too many textures for the
staging buffer to operate in one go.

I had already done this on the wip-bigbsp branch, but this is a better
version: texture uploads are batched as much as possible and the mipmaps
for transparent textures are handled properly.