[qfcc] Expand expressions that result in swizzles
This could become a pessimism (adding 3 swizzles is a fair bit more
expensive than a dot and a scale with [1,1,1] or subtracting two
swizzles is more expensive than a cross product with [1,1,1]), but this
nicely optimizes `S p ~S` where `S = (1 + e23 + e31 + e12)/2` (ie, a
rotation of 120 degrees around the [1,1,1] axis) or its reverse. I don't
know how useful such an optimization really is (the swizzle might be
clearer anyway), and the other 7 axes aren't handled, but it was an
interesting challenge. Also, only the unnormalized version is
recognized.