[gamecode] Check for valid argument access
Checking for too few arguments to sprintf after the fact was "safe" when
the only source of arguments was the hard-coded param slots in progs
memory, but that stopped being the case with the switch to having a
stack and also supporting args from anywhere. This fixes a segfault when
printing a string with at least one format specifier but no arguments.
However, the count is no longer accurate.