[qwaq] Implement inertia tensor calculation
The code is a bit of a mess, but this is my implementation of the
algorithm(s) presented by Steven De Keninck, Martin Roelfs, Leo
Dorst and David Eelbode in their "Clean up your Mesh! Plane and simplex"
paper (https://www.researchgate.net/publication/397490460). Many thanks
to enki (Steven De Keninck) for his help in understanding some of the
trickier bits (in particular, that I was double-dividing by the volume
factorial), and his clarification on equations 5.7 and 5.8 (turned out
to be a typo in the paper). I think the most important thing I got out
of this is the Jacobi algorithm for finding eigenvalues and
eigenvectors (and the Givens rotations used to implement it).