[qfcc] Add dol_super and rename dol_full to dol_sub
The meaning is that for dol_sub, the visited defs are sub-aliases, while
for dol_super, the defs are super-aliasis, compared to the given def.
Unfortunately, this breaks the alias_vec_live test, but it's necessary
for fixing another bug, but this seems to be just a minor error
somewhere in my visit logic.