[qfcc] Use nodes with attached aliases
When a statement uses an alias, and there's a super-alias as an attached
identifier on a node, that node needs to be used instead of creating a
new node. This fixes the brokeninertia test that was due to the chain of
assignments getting lost in the dag. Everything's there in the UD
chains, but the dag code doesn't use them that way and wouldn't be able
to optimize the code anyway. The resulting code still isn't great, but
it's at least correct (and better than before when it worked for other
reasons).