[qfcc] Ensure recycled temps have correct alignment
The recent alignment changes to GA have made it so a bivector temp (6
components) and a struct with a vec4 and two ints (also size 6
(unaligned)) can use the same temp though the struct needs an alignment
of 4 while the bivector doesn't need any alignment. While using a size
of 8 for the struct would fix things for this particular instance, I
think just ensuring the alignment is correct is the better option
because it's entirely possible to have a size 8 object that doesn't need
the same alignment as another size 8 object.