[gamecode] Swap break flag and base register bits
Moving the break flag from bit 15 to 9, and shifting the base register
bits left by 1 allows the opcode to be used directly when indexing into
a jump table that includes break versions of all the instructions. Since
this breaks compatibility with older ruamoko progs, a version bump is
needed, thus I took advantage of the opportunity to fix the design
mistake with the return value size: it's now 12 bits, allowing for 1024
words to be returned, which is probably excessive, but does allow for
R(10) float or R(9) double full multi-vectors.