[build] Force -ffp-contract=off when optimizing
It turns out the default on ARM is `on`, which is what caused the
differences seen in test-quat and test-simd, but I didn't look into it
until I found some qfcc tests failing due to different results from
vdot.f. -ffp-contract=off fixed the qfcc tests, but of course broke
test-quat and test-simd, but they had "are we compiling on arm?" checks
that needed to be modified (made so only non-arm, non-sse (32-bit x86?)
builds have the fudges: ie, arm and x86_64 now get the same results).