[client] Improve the comment about pitch
It was not a "stupid quake bug" as thought (the comment is there in
q1source in gl_rsurf.c), but rather alias models use the convention of
+ve pitch being nose-up, which agrees with aviation. While roll appears
to be correct (+ve roll is starboard wing down), yaw is reversed (+ve
yaw is nose left instead of right). Though, to be fair, maybe the
"stupid bug" was the way rotation was handled in GL Quake. Anyway, a
comment about why the negation is there is certainly better than "stupid
quake bug" :P