[qfcc] Change sizeof to return bytes
This makes it much easier to use with anything expecting to work with
actual data (files, message buffers, etc), though it does cause some fun
with addresses. However, since the goal is compatibility with C, I think
that's a better trade-off.

Note that class instance size is also in terms of bytes now.