[vulkan] Clean out some old vars from draw
Once again on the hunt for the device lost bug with fisheye. It seems to
be something to do with the large sphere mesh and maybe incorrect buffer
updates (Sascha Willems had a similar problem). I know bones aren't
updated safely, but there may be other buffers, and draw does seem to be
connected.