[vulkan] Clean up vertex and index counts for bsp models
index_count was 3 times what it needed to be (I guess I got confused and
thought I was counting triangles), and then index_buffer_size was
rounded up incorrectly. And then, ensure vertex_count and index_count
are never 0 when calculating buffer size because vulkan doesn't like
that, but the empty world model has neither (just the background face at
infinity, and that's drawn using a full screen triangle generated in the
shader).