[qwaq] Add a shader for rendering planetary bodies
It uses an equirectangular map, but with the horizontal axis reversed
from the skymap implementation since it's indented for viewing from the
outside rather than the inside. The shader doesn't get installed
(currently just symlinked from the game directory) nor the embed used,
it's just for making sure the shader gets compiled.